install tests: Added a note reg. shallow clones
I tried to be nice to github and the Internet and not fetch the whole
repository. AFAICS, we only need the tip version anyway.
But git isn't so clever to let us do that.
In the future we might consider making the dependencies submodules of
the PrivacyScore repository and clone recursively.