rescanscanlist: Added a new management command for re-scanning a list
With this change we can trigger a list to be re-scanned.
The Web-UI does not expose a convenient way to do that,
because it requires a form to be submitted.
This, in turn, requires the list to be rendered first in order for us
to get hold of a CSRF token.
However, rendering a large list is currently impossible.